requirements.txt
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
Django==2.2.18
django-cors-headers==3.7.0
django-extensions==3.1.1
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.10
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyDes==2.0.1
Pygments==2.8.1
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
